chore: cleanup, removed unused helpers
