+++ src/management/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './is-privelege-sufficient';
 export * from './map-controller-to-module-name';
+export * from './verify-privileges-for-role';
