+++ src/management/helpers/tests/verify-privileges-for-role.spec.ts
@@ -1,27 +1,7 @@
 import { ModulesEnum, PrivilegesEnum, RolesEnum } from '@src/enums';
 import { UpdatePrivilegesDto } from '@src/management/staff/dto/update-privileges.dto';
 import { getBasePrivileges } from '@src/management/staff/helpers';
-import { rolePrivelegeLimits } from '@src/management/utils/role-privelege-limits';
-import { describe } from 'node:test';
-
-export function verifyPrivilegesForRole(
-  role: RolesEnum,
-  privileges: UpdatePrivilegesDto,
-) {
-  if (role === RolesEnum.ADMIN) return false;
-
-  for (const moduleKey in privileges) {
-    const moduleName = moduleKey as ModulesEnum;
-    const newPrivilege: PrivilegesEnum = privileges[moduleName];
-    const allowedPrivileges = rolePrivelegeLimits[role][moduleName];
-
-    if (!allowedPrivileges.includes(newPrivilege)) {
-      return false;
-    }
-  }
-
-  return true;
-}
+import { verifyPrivilegesForRole } from '@src/management/helpers/verify-privileges-for-role';
 
 describe('verifyPrivilegesForRole', () => {
   it('should return false - there is no possibility to change privileges for the admin role', () => {
