--- src/management/helpers/validate-privileges-values.ts
@@ -1,16 +1,0 @@
-// import { PrivilegesType } from '@src/management/types';
-// import { PrivilegesEnum } from '@src/enums';
-
-// export function validatePrivilegesValues(privileges: PrivilegesType) {
-//   const availablePrivileges = [
-//     PrivilegesEnum.FULL,
-//     PrivilegesEnum.MANAGE,
-//     PrivilegesEnum.READONLY,
-//     PrivilegesEnum.NOT_ALLOWED,
-//   ];
-
-//   if (Object.values(privileges).every((p) => availablePrivileges.includes(p))) {
-//     return true;
-//   }
-//   return false;
-// }
