+++ src/management/staff/helpers/index.ts
@@ -1,2 +1,1 @@
 export * from './get-base-privileges';
-export * from './get-privileges-difference';
